Moved the footerButton to components and Removed the double struct in LessonBoxes and moved it to it's own component
