+++ Badge-IOS App/Badge-IOS App/Components/FooterButton.swift
